Reworked iterface, and call the correct activity for each type of books.

+++ app/src/main/kotlin/com/absreader/data/network/dto/audio_book_library_items/Media.kt
@@ -2,7 +2,7 @@ package com.absreader.data.network.dto.audio_book_library_items
 
 data class Media(
     var coverPath: String?,
-    val duration: Double,
+    val duration: Float,
     val ebookFileFormat: Any,
     val metadata: Metadata,
     val numAudioFiles: Int,
